AP-14 - Cameron James Fixed the find process and current process methods on repositoryService and changed the version numbering of the system.  The branch and processModelVersion for the name use "MAIN" for the initial insert. The version number is taken from the UI on insertion and placed on the processModelVersion.  Also added the wsdl back to the web app, wouldn't work without it and removed unused libraries in the extras folder.

+++ Apromore-Core/apromore-service/src/main/java/org/apromore/service/ProcessService.java
@@ -94,7 +94,7 @@ public interface ProcessService {
      * @return
      * @throws ExceptionDao
      */
-    ProcessModelVersion addProcessModelVersion(ProcessBranch branch, String rootFragmentVersionUri, int versionNumber, String versionName,
+    ProcessModelVersion addProcessModelVersion(ProcessBranch branch, String rootFragmentVersionUri, Double versionNumber, String versionName,
             int numVertices, int numEdges) throws ExceptionDao;
